requirements.txt
Package                       Version
----------------------------- -------
aiohttp                       3.8.4
aiosignal                     1.3.1
asgiref                       3.7.2
async-timeout                 4.0.2
attrs                         23.1.0
charset-normalizer            3.1.0
Django                        4.2.2
django-cors-headers           4.1.0
djangorestframework           3.14.0
djangorestframework-simplejwt 5.2.2
frozenlist                    1.3.3
idna                          3.4
multidict                     6.0.4
pip                           22.3.1
PyJWT                         2.7.0
pytz                          2023.3
setuptools                    66.1.1
sqlparse                      0.4.4
typing_extensions             4.6.3
tzdata                        2023.3
wheel                         0.38.4
yarl                          1.9.2
